Refactor game state management and enhance card selection logic
- Introduced new handler functions for managing blind selections, consumable usage, and shop interactions.
- Refactored card selection and scoring logic to improve clarity and maintainability.
- Updated utility functions for random card generation and pack management to streamline gameplay.
- Enhanced event handling to ensure consistent behavior during game phases and interactions.
