+++ src/app/daily-card-game/domain/game/handlers.ts
@@ -13,36 +13,13 @@ import type { RoundState } from '@/app/daily-card-game/domain/round/types'
 
 import { dealCardsFromDrawPile, getHand } from './card-registry-utils'
 import { HAND_SIZE } from './constants'
-import { calculateInterest } from './reduce-game'
-import { collectEffects, getBlindDefinition } from './utils'
+import { calculateInterest, collectEffects, getBlindDefinition } from './utils'
 
 import type { GamePlayState, GameState } from './types'
 import type { Draft } from 'immer'
 
 type HandEndOutcome = 'gameOver' | 'blindRewards' | 'continue'
 
-/**
- * Gets the probability multiplier based on game state.
- * e.g., "Oops! All 6s" joker doubles all probabilities.
- */
-function getProbabilityMultiplier(game: GameState): number {
-  const hasOopsAll6s = game.jokers.some(j => j.jokerId === 'oopsAll6s')
-  return hasOopsAll6s ? 2 : 1
-}
-
-/**
- * Checks if a lucky roll succeeds.
- * @param baseChance - The base chance as "1 in X" (e.g., 5 means 1 in 5 = 20%)
- * @param seed - Seed string for deterministic random
- * @param game - Game state to check for probability modifiers
- */
-function checkLuckyRoll(baseChance: number, seed: string, game: GameState): boolean {
-  const multiplier = getProbabilityMultiplier(game)
-  // Base probability is 1/baseChance, multiplied by the modifier
-  const probability = Math.min(1, (1 / baseChance) * multiplier)
-  const roll = getRandomFloatWithSeed(seed)
-  return roll < probability
-}
 
 function getCurrentRound(draft: Draft<GameState>): RoundState {
   return draft.rounds[draft.roundIndex]
@@ -187,138 +164,4 @@ export function handleHandScoringEnd(draft: Draft<GameState>, event: GameEvent) 
   dealCardsFromDrawPile(draft as unknown as GameState, cardsNeeded)
   draft.gamePhase = 'gameplay'
   resetScoreForNextHand(draft.gamePlayState)
-}
-
-function scoreCardOnce(
-  draft: GameState,
-  card: (typeof draft.gamePlayState.cardsToScore)[number],
-  event: GameEvent
-) {
-  const gamePlayState = draft.gamePlayState
-
-  let cardChips = playingCards[card.playingCardId].baseChips
-  let cardMult = 0
-
-  // Enchantments (additive)
-  if (card.flags.enchantment === 'bonus') cardChips += 30
-  if (card.flags.enchantment === 'mult') cardMult += 5
-
-  // Editions (additive)
-  if (card.flags.edition === 'foil') cardChips += 50
-  if (card.flags.edition === 'holographic') cardMult += 10
-
-  if (cardChips > 0) {
-    draft.gamePlayState.scoringEvents.push({
-      id: uuid(),
-      type: 'chips',
-      value: cardChips,
-      source: playingCards[card.playingCardId].value,
-    })
-  }
-
-  if (cardMult > 0) {
-    draft.gamePlayState.scoringEvents.push({
-      id: uuid(),
-      type: 'mult',
-      value: cardMult,
-      source: playingCards[card.playingCardId].value,
-    })
-  }
-
-  gamePlayState.score = {
-    chips: gamePlayState.score.chips + cardChips,
-    mult: gamePlayState.score.mult + cardMult,
-  }
-
-  // Polychrome edition: X1.5 Mult (multiplicative, applied after additive bonuses)
-  if (card.flags.edition === 'polychrome') {
-    gamePlayState.score.mult *= 1.5
-    draft.gamePlayState.scoringEvents.push({
-      id: uuid(),
-      type: 'mult',
-      operator: 'x',
-      value: 1.5,
-      source: playingCards[card.playingCardId].value,
-    })
-  }
-
-  // Gold seal: earn $3 immediately when scored
-  if (card.flags.seal === 'gold') {
-    draft.money += 3
-  }
-
-  // Lucky enchantment: 1 in 5 chance for +20 Mult, 1 in 15 chance for $20
-  // Both effects roll separately and can both trigger
-  if (card.flags.enchantment === 'lucky') {
-    const multSeed = buildSeedString([
-      draft.gameSeed,
-      card.id,
-      'lucky-mult',
-      String(draft.handsPlayed),
-    ])
-    const moneySeed = buildSeedString([
-      draft.gameSeed,
-      card.id,
-      'lucky-money',
-      String(draft.handsPlayed),
-    ])
-
-    if (checkLuckyRoll(5, multSeed, draft)) {
-      draft.gamePlayState.scoringEvents.push({
-        id: uuid(),
-        type: 'mult',
-        value: 20,
-        source: 'Lucky',
-      })
-      gamePlayState.score.mult += 20
-    }
-
-    if (checkLuckyRoll(15, moneySeed, draft)) {
-      draft.money += 20
-    }
-  }
-
-  const currentBlind = getInProgressBlind(draft as unknown as GameState)
-  if (!currentBlind) return
-
-  const playedCards = draft.gamePlayState.selectedHand?.[1]
-  const ctx: EffectContext = {
-    event,
-    game: draft as unknown as GameState,
-    score: gamePlayState.score,
-    playedCards,
-    scoredCards: [card],
-    round: draft.rounds[draft.roundIndex],
-    bossBlind: draft.rounds[draft.roundIndex].bossBlind,
-    jokers: draft.jokers,
-    vouchers: draft.vouchers,
-    tags: draft.tags,
-  }
-  dispatchEffects(event, ctx, collectEffects(ctx.game))
-}
-
-export function handleCardScored(draft: GameState, event: GameEvent) {
-  const gamePlayState = draft.gamePlayState
-  const currentCardToScore = gamePlayState.cardsToScore.shift()
-  if (!currentCardToScore) return
-  const scoredCardId = currentCardToScore.id
-  const hasRedSeal = currentCardToScore.flags.seal === 'red'
-
-  // Score the card (and score again if it has a red seal)
-  scoreCardOnce(draft, currentCardToScore, event)
-  if (hasRedSeal) {
-    scoreCardOnce(draft, currentCardToScore, event)
-  }
-
-  // remove card from selection & hand UI (only once, regardless of red seal)
-  gamePlayState.selectedCardIds = gamePlayState.selectedCardIds.filter(id => id !== scoredCardId)
-  if (gamePlayState.selectedHand) {
-    gamePlayState.selectedHand = [
-      gamePlayState.selectedHand[0],
-      gamePlayState.selectedHand[1].filter(card => card.id !== scoredCardId),
-    ]
-  }
-
-  // remove scored card from hand
-  gamePlayState.handIds = gamePlayState.handIds.filter(id => id !== scoredCardId)
 }
