+++ src/app/daily-card-game/domain/game/utils.ts
@@ -7,8 +7,9 @@ import {
   isTarotCardState,
 } from '@/app/daily-card-game/domain/consumable/utils'
 import { dispatchEffects } from '@/app/daily-card-game/domain/events/dispatch-effects'
-import type { Effect, GameEvent } from '@/app/daily-card-game/domain/events/types'
+import type { Effect, EffectContext, GameEvent } from '@/app/daily-card-game/domain/events/types'
 import { jokers } from '@/app/daily-card-game/domain/joker/jokers'
+import type { JokerState } from '@/app/daily-card-game/domain/joker/types'
 import type { PlayingCardState } from '@/app/daily-card-game/domain/playing-card/types'
 import { mulberry32, xmur3 } from '@/app/daily-card-game/domain/randomness'
 import { bigBlind, getInProgressBlind, smallBlind } from '@/app/daily-card-game/domain/round/blinds'
@@ -19,7 +20,10 @@ import type {
   RoundState,
 } from '@/app/daily-card-game/domain/round/types'
 import { implementedTags as tags } from '@/app/daily-card-game/domain/tag/tags'
+import { getRandomTag } from '@/app/daily-card-game/domain/tag/utils'
 import { vouchers } from '@/app/daily-card-game/domain/voucher/vouchers'
+
+import { INTEREST_CALCULATION_FACTOR } from './constants'
 
 import type { GameState } from './types'
 
@@ -97,48 +101,50 @@ export function shuffleCardIds({
   return shuffleArray(cardIds, rng)
 }
 
-export function useBuyableCelestialCard(draft: Draft<GameState>): void {
-  const celestialCardId = draft.shopState.selectedCardId
-  if (!celestialCardId) return
-  const celestialCard = draft.shopState.cardsForSale.find(card => card.card.id === celestialCardId)
-  if (!celestialCard) return
-  if (isCelestialCardState(celestialCard.card)) {
-    // Add to consumablesUsed so The Fool and similar cards can reference it
-    draft.consumablesUsed.push(celestialCard.card)
-    // Important: `dispatchEffects` matches by `event.type`. When buying + using in the shop,
-    // the triggering event is typically `SHOP_BUY_AND_USE_CARD`, but the card's effects are
-    // keyed off `CELESTIAL_CARD_USED`.
-    const usedEvent: GameEvent = { type: 'CELESTIAL_CARD_USED' }
-    dispatchEffects(
-      usedEvent,
-      {
-        event: usedEvent,
-        game: draft,
-        score: draft.gamePlayState.score,
-        playedCards: [],
-        round: draft.rounds[draft.roundIndex],
-        bossBlind: draft.rounds[draft.roundIndex].bossBlind,
-        jokers: draft.jokers,
-        vouchers: draft.vouchers,
-        tags: draft.tags,
-      },
-      celestialCards[celestialCard.card.handId].effects
-    )
+
+export function removeJoker(draft: GameState, event: GameEvent, selectedJoker: JokerState) {
+  draft.gamePlayState.selectedJokerId = undefined
+  const ctx: EffectContext = {
+    event,
+    game: draft as unknown as GameState,
+    score: draft.gamePlayState.score,
+    playedCards: [],
+    round: draft.rounds[draft.roundIndex],
+    bossBlind: draft.rounds[draft.roundIndex].bossBlind,
+    jokers: draft.jokers,
+    vouchers: draft.vouchers,
+    tags: draft.tags,
   }
+  // Collect effects *before* removing the joker so "on sold/removed" effects that live on the
+  // removed joker itself still get a chance to run. Then dispatch *after* removal so effects
+  // can observe the post-removal game state.
+  const effectsBeforeRemoval = collectEffects(ctx.game)
+  draft.jokers = draft.jokers.filter(joker => joker.id !== selectedJoker.id)
+  ctx.jokers = draft.jokers
+  dispatchEffects(event, ctx, effectsBeforeRemoval)
 }
 
-export function useConsumableCelestialCard(draft: Draft<GameState>, event: GameEvent): void {
-  const celestialCard = draft.gamePlayState.selectedConsumable
-  if (!celestialCard) return
-  if (celestialCard.consumableType !== 'celestialCard') return
-  draft.consumablesUsed.push(celestialCard)
-  draft.gamePlayState.selectedConsumable = undefined
-  // remove tarot card from consumables
-  draft.consumables = draft.consumables.filter(consumable => consumable.id !== celestialCard.id)
+export function calculateInterest(draft: GameState): number {
+  const currentMoney = draft.money
+  const maxInterest = draft.maxInterest
+  const interestCalculation = Math.floor(currentMoney / INTEREST_CALCULATION_FACTOR)
+  return Math.min(interestCalculation, maxInterest)
+}
 
-  dispatchEffects(
-    event,
-    {
+export function populateTags(draft: GameState): void {
+  const bigBlindTag = getRandomTag(draft)
+  const smallBlindTag = getRandomTag(draft)
+  draft.rounds[draft.roundIndex].bigBlind.tag = bigBlindTag
+  draft.rounds[draft.roundIndex].smallBlind.tag = smallBlindTag
+}
+
+export function getEffectContext(
+  draft: GameState,
+  event: GameEvent,
+  override?: Partial<EffectContext>
+): EffectContext {
+  return {
+    ...{
       event,
       game: draft,
       score: draft.gamePlayState.score,
@@ -149,61 +155,6 @@ export function useConsumableCelestialCard(draft: Draft<GameState>, event: GameE
       vouchers: draft.vouchers,
       tags: draft.tags,
     },
-    celestialCards[celestialCard.handId].effects
-  )
-}
-
-export function useBuyableTarotCard(draft: Draft<GameState>): void {
-  const tarotCardId = draft.shopState.selectedCardId
-  if (!tarotCardId) return
-  const tarotCard = draft.shopState.cardsForSale.find(card => card.card.id === tarotCardId)
-  if (!tarotCard) return
-  if (isTarotCardState(tarotCard.card)) {
-    // Add to consumablesUsed so The Fool and similar cards can reference it
-    draft.consumablesUsed.push(tarotCard.card)
-    // See note in `useBuyableCelestialCard`: buy+use triggers a shop event, but tarot effects
-    // are keyed off `TAROT_CARD_USED`.
-    const usedEvent: GameEvent = { type: 'TAROT_CARD_USED' }
-    dispatchEffects(
-      usedEvent,
-      {
-        event: usedEvent,
-        game: draft,
-        score: draft.gamePlayState.score,
-        playedCards: [],
-        round: draft.rounds[draft.roundIndex],
-        bossBlind: draft.rounds[draft.roundIndex].bossBlind,
-        jokers: draft.jokers,
-        vouchers: draft.vouchers,
-        tags: draft.tags,
-      },
-      tarotCards[tarotCard.card.tarotType].effects
-    )
+    ...override,
   }
-}
-
-export function useConsumableTarotCard(draft: Draft<GameState>, event: GameEvent): void {
-  const tarotCard = draft.gamePlayState.selectedConsumable
-  if (!tarotCard) return
-  if (tarotCard.consumableType !== 'tarotCard') return
-  if (tarotCard.tarotType === 'notImplemented') return
-  draft.consumablesUsed.push(tarotCard)
-  draft.gamePlayState.selectedConsumable = undefined
-  draft.consumables = draft.consumables.filter(consumable => consumable.id !== tarotCard.id)
-
-  dispatchEffects(
-    event,
-    {
-      event,
-      game: draft as unknown as GameState,
-      score: draft.gamePlayState.score,
-      playedCards: [],
-      round: draft.rounds[draft.roundIndex],
-      bossBlind: draft.rounds[draft.roundIndex].bossBlind,
-      jokers: draft.jokers,
-      vouchers: draft.vouchers,
-      tags: draft.tags,
-    },
-    tarotCards[tarotCard.tarotType].effects
-  )
 }
